--- web/apps/photos/src/components/Collections/CollectionShare/publicShare/EnablePublicShareOptions.tsx
@@ -1,103 +1,0 @@
-import {
-    MenuItemDivider,
-    MenuItemGroup,
-    MenuSectionTitle,
-} from "@/base/components/Menu";
-import type { Collection, PublicURL } from "@/media/collection";
-import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
-import DownloadSharp from "@mui/icons-material/DownloadSharp";
-import LinkIcon from "@mui/icons-material/Link";
-import PublicIcon from "@mui/icons-material/Public";
-import { Stack, Typography } from "@mui/material";
-import { t } from "i18next";
-import { GalleryContext } from "pages/gallery";
-import { useContext, useState } from "react";
-import {
-    createShareableURL,
-    updateShareableURL,
-} from "services/collectionService";
-import { handleSharingErrors } from "utils/error/ui";
-interface Iprops {
-    collection: Collection;
-    setPublicShareProp: (value: PublicURL) => void;
-    onLinkCreated: () => void;
-}
-
-export default function EnablePublicShareOptions({
-    collection,
-    setPublicShareProp,
-    onLinkCreated,
-}: Iprops) {
-    const galleryContext = useContext(GalleryContext);
-    const [sharableLinkError, setSharableLinkError] = useState(null);
-
-    const createSharableURLHelper = async () => {
-        try {
-            setSharableLinkError(null);
-            galleryContext.setBlockingLoad(true);
-            const publicURL = await createShareableURL(collection);
-            setPublicShareProp(publicURL);
-            onLinkCreated();
-            galleryContext.syncWithRemote(false, true);
-        } catch (e) {
-            const errorMessage = handleSharingErrors(e);
-            setSharableLinkError(errorMessage);
-        } finally {
-            galleryContext.setBlockingLoad(false);
-        }
-    };
-
-    const createCollectPhotoShareableURLHelper = async () => {
-        try {
-            setSharableLinkError(null);
-            galleryContext.setBlockingLoad(true);
-            const publicURL = await createShareableURL(collection);
-            await updateShareableURL({
-                collectionID: collection.id,
-                enableCollect: true,
-            });
-            setPublicShareProp(publicURL);
-            onLinkCreated();
-            galleryContext.syncWithRemote(false, true);
-        } catch (e) {
-            const errorMessage = handleSharingErrors(e);
-            setSharableLinkError(errorMessage);
-        } finally {
-            galleryContext.setBlockingLoad(false);
-        }
-    };
-
-    return (
-        <Stack>
-            <MenuSectionTitle
-                title={t("LINK_SHARE_TITLE")}
-                icon={<PublicIcon />}
-            />
-            <MenuItemGroup>
-                <EnteMenuItem
-                    label={t("CREATE_PUBLIC_SHARING")}
-                    startIcon={<LinkIcon />}
-                    onClick={createSharableURLHelper}
-                />
-                <MenuItemDivider hasIcon />
-                <EnteMenuItem
-                    label={t("COLLECT_PHOTOS")}
-                    startIcon={<DownloadSharp />}
-                    onClick={createCollectPhotoShareableURLHelper}
-                />
-            </MenuItemGroup>
-            {sharableLinkError && (
-                <Typography
-                    textAlign={"center"}
-                    variant="small"
-                    sx={{
-                        color: (theme) => theme.colors.danger.A700,
-                        mt: 0.5,
-                    }}
-                >
-                    {sharableLinkError}
-                </Typography>
-            )}
-        </Stack>
-    );
-}
